refactor edit profile cuz the layout a bit weird
